Move types to ./types.ts
