+++ packages/js/data/src/plugins/types.ts
@@ -2,6 +2,7 @@
  * Internal dependencies
  */
 import { pluginNames } from './constants';
+import { WPError } from '../types';
 
 export type RecommendedTypes = 'payments';
 
@@ -61,3 +62,21 @@ export type PaypalOnboardingStatus = {
 		onboarded: boolean;
 	};
 };
+
+type PluginsResponse< PluginData > = {
+	data: PluginData;
+	errors: WPError< PluginNames >;
+	success: boolean;
+	message: string;
+} & Response;
+
+export type InstallPluginsResponse = PluginsResponse< {
+	installed: string[];
+	results: Record< string, boolean >;
+	install_time?: Record< string, number >;
+} >;
+
+export type ActivatePluginsResponse = PluginsResponse< {
+	activated: string[];
+	active: string[];
+} >;
